refactor: wishlist - services file structure
