--- src/services/wishlist-services/addToWishlistService.js
@@ -1,10 +1,0 @@
-import axios from "axios";
-
-const addToWishlistService = async (product, encodedToken) =>
-  await axios.post(
-    "/api/user/wishlist",
-    { product },
-    { headers: { authorization: encodedToken } }
-  );
-
-export default addToWishlistService;
