--- src/services/wishlist-services/getWishlistService.js
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-const getWishlistService = async ( encodedToken ) =>
-  await axios.get("/api/user/wishlist", {
-    headers: { authorization: encodedToken },
-  });
-
-export default getWishlistService;
